feat: add composable core export and benchmarks
