+++ packages/sugar-high/lib/presets/index.js
@@ -21,3 +21,5 @@ export * as powershell from './lang/powershell.js'
 export * as dockerfile from './lang/dockerfile.js'
 export * as graphql from './lang/graphql.js'
 export * as hcl from './lang/hcl.js'
+export * as javascript from './lang/javascript.js'
+export * as typescript from './lang/typescript.js'
