+++ packages/sugar-high/lib/presets/lang/typescript.js
@@ -1,0 +1,7 @@
+// @ts-check
+import { tokenize as tokenizeJavaScript } from './javascript-runtime.js'
+
+export const tokenize = (code, options) => tokenizeJavaScript(code, {
+  ...options,
+  typescript: true,
+})
