+++ packages/sugar-high/package.json
@@ -14,6 +14,10 @@
       "types": "./lib/index.d.ts",
       "default": "./lib/index.js"
     },
+    "./core": {
+      "types": "./lib/core.d.ts",
+      "default": "./lib/core.js"
+    },
     "./presets": {
       "types": "./lib/presets/index.d.ts",
       "default": "./lib/presets/index.js"
@@ -30,7 +34,8 @@
   "license": "MIT",
   "scripts": {
     "test": "vitest",
-    "build": "echo 'package requires no build'"
+    "build": "echo 'package requires no build'",
+    "benchmark": "node scripts/benchmark.mjs"
   },
   "devDependencies": {
     "@types/node": "22.12.0",
